Move add-modals to inside the modals dir
To keep all modals in one place, and to used in various locations, not
only the editor (like have happened to other dialogs previously)
