+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/add-widgets-view.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var Backbone = require('backbone');
 var cdb = require('cartodb.js');
-var LoadingView = require('../../components/loading/loading-view');
+var LoadingView = require('../../loading/loading-view');
 var createTuplesItems = require('./create-tuples-items');
 var widgetsTypes = require('./widgets-types');
 var BodyView = require('./body-view');
