+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/add-widgets.tpl
