+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/body-view.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var cdb = require('cartodb.js');
-var createTextLabelsTabPane = require('../../components/tab-pane/create-text-labels-tab-pane');
+var createTextLabelsTabPane = require('../../tab-pane/create-text-labels-tab-pane');
 var tabPaneTemplate = require('./tab-pane-template.tpl');
 
 /**
