+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/category/category-options-view.js
