+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/create-tuples-items.js
