+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/formula/formula-option.tpl
