+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/histogram/histogram-option-model.js
