+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/layer-selector-view.js
