+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/tab-pane-template.tpl
