+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/time-series/time-series-option.tpl
