+++ lib/assets/javascripts/cartodb3/components/modals/add-widgets/widgets-types.js
