+++ lib/assets/javascripts/cartodb3/editor/map/map-view.js
@@ -3,11 +3,11 @@ var Backbone = require('backbone');
 var cdb = require('cartodb-deep-insights.js');
 var EditorWidgetsView = require('../widgets/widgets-view');
 var createTextLabelsTabPane = require('../../components/tab-pane/create-text-labels-tab-pane');
+var AddWidgetsView = require('../../components/modals/add-widgets/add-widgets-view');
 var StackLayoutView = require('../../components/stack-layout/stack-layout-view');
 var WidgetsFormContentView = require('../widgets/widgets-form/widgets-form-content-view');
 var Template = require('./template.tpl');
 var TabPaneTemplate = require('./tab-pane-template.tpl');
-var AddWidgetsView = require('../add-widgets/add-widgets-view');
 
 module.exports = cdb.core.View.extend({
   initialize: function (opts) {
