+++ lib/assets/test/spec/cartodb3/components/modals/add-widgets/add-widgets-view.spec.js
@@ -1,10 +1,10 @@
 var cdb = require('cartodb.js');
 var Backbone = require('backbone');
-var ConfigModel = require('../../../../../javascripts/cartodb3/data/config-model');
-var LayerDefinitionsCollection = require('../../../../../javascripts/cartodb3/data/layer-definitions-collection');
-var AddWidgetsView = require('../../../../../javascripts/cartodb3/editor/add-widgets/add-widgets-view');
+var ConfigModel = require('../../../../../../javascripts/cartodb3/data/config-model');
+var LayerDefinitionsCollection = require('../../../../../../javascripts/cartodb3/data/layer-definitions-collection');
+var AddWidgetsView = require('../../../../../../javascripts/cartodb3/components/modals/add-widgets/add-widgets-view');
 
-describe('editor/add-widgets/add-widgets-view', function () {
+describe('components/modals/add-widgets/add-widgets-view', function () {
   var FETCHING_TITLE = 'fetching-';
 
   beforeEach(function () {
