+++ lib/assets/test/spec/cartodb3/components/modals/add-widgets/body-view.spec.js
@@ -1,9 +1,9 @@
 var cdb = require('cartodb.js');
 var Backbone = require('backbone');
-var BodyView = require('../../../../../javascripts/cartodb3/editor/add-widgets/body-view');
-var widgetsTypes = require('../../../../../javascripts/cartodb3/editor/add-widgets/widgets-types');
+var BodyView = require('../../../../../../javascripts/cartodb3/components/modals/add-widgets/body-view');
+var widgetsTypes = require('../../../../../../javascripts/cartodb3/components/modals/add-widgets/widgets-types');
 
-describe('editor/add-widgets/body-view', function () {
+describe('components/modals/add-widgets/body-view', function () {
   beforeEach(function () {
     var layerDefinitionModel = new cdb.core.Model();
     layerDefinitionModel.getName = function () { return 'layer'; };
