+++ lib/assets/test/spec/cartodb3/components/modals/add-widgets/create-tuples-items.spec.js
@@ -1,10 +1,10 @@
 var _ = require('underscore');
 var Backbone = require('backbone');
-var ConfigModel = require('../../../../../javascripts/cartodb3/data/config-model');
-var LayerDefinitionsCollection = require('../../../../../javascripts/cartodb3/data/layer-definitions-collection');
-var createTuplesItems = require('../../../../../javascripts/cartodb3/editor/add-widgets/create-tuples-items');
+var ConfigModel = require('../../../../../../javascripts/cartodb3/data/config-model');
+var LayerDefinitionsCollection = require('../../../../../../javascripts/cartodb3/data/layer-definitions-collection');
+var createTuplesItems = require('../../../../../../javascripts/cartodb3/components/modals/add-widgets/create-tuples-items');
 
-describe('editor/add-widgets/create-tuples-items', function () {
+describe('components/modals/add-widgets/create-tuples-items', function () {
   beforeEach(function () {
     var configModel = new ConfigModel({
       base_url: '/u/pepe'
