+++ lib/assets/test/spec/cartodb3/components/modals/add-widgets/layer-selector-view.spec.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js');
-var LayerSelectorView = require('../../../../../javascripts/cartodb3/editor/add-widgets/layer-selector-view');
+var LayerSelectorView = require('../../../../../../javascripts/cartodb3/components/modals/add-widgets/layer-selector-view');
 
-describe('editor/add-widgets/layer-selector-view', function () {
+describe('components/modals/add-widgets/layer-selector-view', function () {
   beforeEach(function () {
     this.columnModel1 = new cdb.core.Model({
       name: 'col',
