+++ lib/assets/test/spec/cartodb3/components/modals/add-widgets/widgets-types.spec.js
@@ -1,9 +1,9 @@
 var _ = require('underscore');
 var Backbone = require('backbone');
-var widgetsTypes = require('../../../../../javascripts/cartodb3/editor/add-widgets/widgets-types');
-var TimeSeriesNoneOptionModel = require('../../../../../javascripts/cartodb3/editor/add-widgets/time-series/time-series-none-option-model');
+var widgetsTypes = require('../../../../../../javascripts/cartodb3/components/modals/add-widgets/widgets-types');
+var TimeSeriesNoneOptionModel = require('../../../../../../javascripts/cartodb3/components/modals/add-widgets/time-series/time-series-none-option-model');
 
-describe('editor/add-widgets/widgets-types', function () {
+describe('components/modals/add-widgets/widgets-types', function () {
   describe('all items', function () {
     it('should have a type prop', function () {
       widgetsTypes.forEach(function (d) {
